feat: AI 채팅 UI 구현
